Timer sensor updated to uint16_t

+++ ithofanComponent/StatusMessage.h
@@ -11,5 +11,5 @@ public:
     StatusMessage(std::vector<uint8_t> messageBytes);
     bool valid() const;
     FanStatus getFanStatus() const;
-    uint8_t getRemainingTime() const;
+    uint16_t getRemainingTime() const;
 };
